+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## [v1.18.1](https://github.com/vazco/uniforms/tree/v1.18.1) (2017-05-28)
 - **Changed:** Optimized `BaseField` props calculation.
 - **Changed:** Switched to `prop-types`.
 - **Fixed:** `Bridge.check` error message.